requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
emails==0.6
aiosmtplib==3.0.1
email-validator==2.1.0.post1
Jinja2==3.1.3
uvicorn[standard]==0.27.0